requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
lxml==4.6.3
    # via python-docx
pymupdf==1.18.19
    # via -r requirements.in
pypdf2==1.26.0
    # via -r requirements.in
python-docx==0.8.11
    # via -r requirements.in
